[GATE-40] Automatically wiping existing pseudo roster items on new create. [GATE-46] Logistical error causes buddy list to be wiped on migration from py transports.  Disabled deletes for now. [GATE-45] Pseudo roster is now cleaned up.
git-svn-id: http://svn.igniterealtime.org/svn/repos/wildfire/trunk/src/plugins/gateway@5423 b35dd754-fafc-0310-a699-88a17e54d16e

+++ src/database/upgrade/1/gateway_sqlserver.sql
@@ -3,7 +3,7 @@ ALTER TABLE gatewayRegistration ADD nickname NVARCHAR(255);
 
 /* Add pseudo roster table */
 CREATE TABLE gatewayPseudoRoster (
-   registrationID    INTEGER        NOT NULL,
+   registrationID    BIGINT         NOT NULL,
    username          NVARCHAR(255)  NOT NULL,
    nickname          NVARCHAR(255),
    groups            NVARCHAR(255)
